Create README.md
Application to serve a sort of reading list for the users. Users are going to be able to use the application to put together a list of books that they would like to read. 
User can add in a new book by using a form that is it always going to be display down at the bottom of the screen. User should be able to type in something like a new book 
and then click on "Submit" button, after clicking on it the text they entered, will be passed as a new card. Once a user has assembled a reading list of books, were then going 
to allow them to edit or delete these books. User can delete a book by clicking on the "x" on right top corner side of every one of these cards. In case the user would like 
to change the title of a book, next to the "x" there is a pencil icon that toggles the edit mode for one individual book and by pressing the "save" button updates it.